requirements.txt
Django==3.2.*
djangorestframework==3.12.2
django-cors-headers==3.3.0
django-cors-middleware==1.3.1
bcrypt==4.0.1
PyJWT==1.7.1
cryptography==41.0.1
django-ratelimit==4.0.0
